Move geocoding to a component.
