+++ app/components/geocoding/__init__.py
@@ -1,0 +1,6 @@
+from . import routes, schemas
+
+__all__ = [
+    "routes",
+    "schemas",
+]
