+++ app/components/geocoding/enums.py
@@ -1,0 +1,6 @@
+from app.schemas import BasicEnum
+
+
+class GeocoderEnum(BasicEnum):
+    gmaps = "gmaps"
+    osm = "osm"
