+++ app/schemas/__init__.py
@@ -1,3 +1,2 @@
 from .enums import *
-from .geocoding import *
 from .roles import UserRole
