+++ app/schemas/enums.py
@@ -43,6 +43,3 @@ class HospitalEnum(BasicEnum):
     no_data = "no_data"
 
 
-class GeocoderEnum(BasicEnum):
-    gmaps = "gmaps"
-    osm = "osm"
